refactor(calendar): isolate misc capability module
